+++ src/Be.Vlaanderen.Basisregisters.Api/ETag/ContentWithETagResult.cs
@@ -4,7 +4,6 @@ namespace Be.Vlaanderen.Basisregisters.Api.ETag
     using Microsoft.AspNetCore.Mvc;
     using Microsoft.Net.Http.Headers;
 
-    [Obsolete("Phase out: switch to using LastObservedPosition", false)]
     public class ContentWithETagResult : ContentResult
     {
         private readonly string _etag;